refactor model to use Double instead ofg BigDecimal

+++ src/main/java/com/vape/data/streaming/model/SensorDataPointModel.java
@@ -4,7 +4,6 @@ import lombok.*;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -19,5 +18,5 @@ public class SensorDataPointModel {
     private String sensorId;
     private String sensorHubId;
     private String timestamp;
-    private List<BigDecimal> data;
+    private List<Double> data;
 }
